Resolve PR comment
